fixing error
