+++ rust_http_server/src/http/method.rs
@@ -1,0 +1,11 @@
+pub enum Method {
+            GET,
+            DELETE,
+            POST,
+            PUT,
+            HEAD,
+            CONNECT,
+            OPTIONS,
+            TRACE,
+            PATCH
+        }
