+++ rust_http_server/src/http/mod.rs
@@ -1,0 +1,5 @@
+pub use request::Request;
+pub use method::Method;
+
+pub mod request;
+pub mod method;
