+++ rust_http_server/src/http/request.rs
@@ -1,0 +1,6 @@
+ use super::method::Method;
+        pub struct Request{
+            path: String,
+            query_string: Option<String>,
+            method: Method,
+        }
